Restructure (review feedback)
